+++ core/model/meme.py
@@ -1,6 +1,6 @@
 import datetime
 
-from sqlalchemy import Column, BigInteger, String, ForeignKey, Index, DateTime
+from sqlalchemy import Column, BigInteger, String, ForeignKey, Index, DateTime, Integer
 
 from . import Base
 
@@ -10,7 +10,7 @@ class Meme(Base):
 
     phrase = Column(String, primary_key=True, index=True)
     creator_id = Column(BigInteger, ForeignKey('user.user_id'))
-    uses = Column(BigInteger, default=-1)
+    uses = Column(Integer, default=-1)
     image = Column(String)
     created_at = Column(DateTime, default=datetime.datetime.now(datetime.timezone.utc))
     content = Column(String)
